requirements.txt
Pillow==9.2.0
pydantic==1.9.1
requests==2.25.1
